Editing publication works

+++ coders-space-frontend/src/app/publication/publication.component.ts
@@ -1,43 +1,45 @@
-import { Component, EventEmitter, Input, Output, } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { Publication } from '../interfaces/publication';
 import { PublicationService } from '../services/publication.service';
 import { EditPublicationComponent } from '../edit-publication/edit-publication.component';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-publication',
   standalone: true,
-  imports: [EditPublicationComponent],
+  imports: [EditPublicationComponent, CommonModule],
   templateUrl: './publication.component.html',
-  styleUrl: './publication.component.css'
+  styleUrls: ['./publication.component.css']
 })
 export class PublicationComponent {
   @Input() publication!: Publication;
   @Output() publicationsChanged = new EventEmitter<void>();
   isModalVisible: boolean = false;
 
+  constructor(private publicationService: PublicationService) { }
+
   editPublication(): void {
     this.isModalVisible = true;
-    console.log('Edit publication');
+    console.log('Opening edit publication modal');
   }
 
   closeEditPublication(): void {
     this.isModalVisible = false;
-    console.log('Close edit publication')
+    console.log('Closing edit publication modal');
   }
-
-  constructor(private publicationService: PublicationService) { }
 
   public removePublication() {
-    console.log('Publication will be deleted');
-    this.publicationService.deletePublication(this.publication).subscribe({
-      next: (response: any) => {
-        console.log('Publication deleted', response);
-        this.publicationsChanged.emit();  // Emit event to parent
-      },
-      error: (error: any) => {
-        console.error('There was an error!', error);
-      },
-    });
+    if (confirm('Are you sure you want to delete this publication?')) {
+      this.publicationService.deletePublication(this.publication).subscribe({
+        next: (response: any) => {
+          console.log('Publication deleted', response);
+          this.publicationsChanged.emit();
+        },
+        error: (error: any) => {
+          console.error('There was an error!', error);
+        },
+      });
+    }
   }
-
 }
+
